gcexport.py works in python2 and 3 now
